+++ pacli/config.py
@@ -1,5 +1,7 @@
+from appdirs import user_config_dir
+import logging
 import configparser
-import sys
+import os, sys
 from .default_conf import default_conf
 
 def write_default_config(conf_file=None):
@@ -39,3 +41,32 @@ def read_conf(conf_file):
         settings["testnet"] = True
 
     return settings
+
+
+conf_dir = user_config_dir("pacli")
+conf_file = os.path.join(conf_dir, "pacli.conf")
+logfile = os.path.join(conf_dir, "pacli.log")
+
+
+def load_conf():
+    '''load user configuration'''
+
+    class Settings:
+        pass
+
+    settings = read_conf(conf_file)
+
+    for key in settings:
+        setattr(Settings, key, settings[key])
+
+    logging.basicConfig(filename=logfile, level=logging.getLevelName(Settings.loglevel))
+    logging.basicConfig(level=logging.getLevelName(Settings.loglevel),
+                        format="%(asctime)s %(levelname)s %(message)s")
+
+    logging.debug("logging initialized")
+
+    return Settings
+
+
+Settings = load_conf()
+
